+++ contracts/src/models/map.cairo
@@ -4,7 +4,7 @@ use darkshuffle::models::battle::{CreatureType};
 #[dojo::model]
 pub struct Map {
     #[key]
-    game_id: u128,
+    game_id: u64,
     #[key]
     level: u8,
     seed: u128
